Adding nested property categories
