+++ Objects/Converters/ConverterNavisworks/ConverterNavisworks/ConverterNavisworksShared.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>ConverterNavisworksShared</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)ConverterNavisworks.Properties.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ConverterNavisworks.ToNative.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ConverterNavisworks.ToSpeckle.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ConverterNavisworks.cs" />
